Add the accountListId to the finish page URL
